Se crean las clases documento y economicDocument, Se busca clientes y productos por ID, se borran clases inutiles (firstSerlet,ModeServlet, y Empleado)

+++ src/com/divingWeb/conexionDAO/ClientDAO.java
@@ -40,7 +40,7 @@ public class ClientDAO extends ConexionDAO {
 		try 
         { 
             iniciaOperacion(); 
-            cl = (Cliente)sesion.get(Cliente.class, idCliente);
+            cl = (Cliente)sesion.get(Cliente.class, (long)idCliente);
             tx.commit(); 
         } catch (HibernateException he) 
         {
